requirements.txt
WTForms~=3.0.1
Flask~=2.0.2
mysql-connector-python~=8.0.28